Render super category content
Don't filter out super categories as they may have intro content to be
rendered even if they don't contain tools.

Instead, fetch the tools for the category only if it's not a super
category, then render as normal.

Also change up how whitespace around the intro content is managed,
because it can be preceded by a newline if the description field is
frontmatter is present in the markdown file.