If no url detected for a tool, default to the toolkit entry itself
It is becoming increasingly common to add a toolkit entry which is a
meta-entry, i.e., a list of related online resources or a guide in-situ.
In this case there will be no external URL to link to (in a category
list for example), so we will simply link to the toolkit entry itself in
this case.

Also attempted to improve the log output for these special cases and in
cases where the URL is overridden in json.